Create new file for LintErrors
