+++ lib/sample_code.rb
@@ -5,7 +5,7 @@ puts 'hello people ';
 end;
 
 
-i = 10; i -= 1; i = 43;
+i = 10; i -= 1; i = 43; 
 puts 'Hello world! Hello world! Hello world! Hello world! Hello world! Hello world! Hello world! Hello world! ';
